Replaced c modules with c++ modules.
